Fixed a bug in RemoveDeadSymbols in which if multiple function's DefExprs shared the same DefStmt, they would all be extracted if one was dead.  Now just the DefExpr is extracted if there are multiple DefExprs.  If there is a single DefExpr, the DefStmt is extracted.
If --useNewConstructor is on, I now run call_info on the DefExpr to
find out about the defaultConstructor.  This causes one test to fail
so I only do it if the flag is on.

This checkin gets the named parameter constructor working with most
record tests.  There are still a few failures with this flag so it is
not the default.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2555 3a8e244f-b0f2-452b-bcba-4c88e055c3ca