Completed decoupling methods from classes.  Methods, including getters, setters, and constructors, can now stay alive even if the class dies.  This applies to records, unions, etc.
Scopes of functions are now really only used for code generation and
for computing the visible functions list.

Secondary methods are now in the scopes where they are defined.

Cleaned up code for building constructor bodies.

Code generation of a structural type no longer results in code
generation of its methods or constructors.

Write functions are only inserted if the type is used.  If the type is
dead, a write function is not created.

Fixes to gross sequence code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2568 3a8e244f-b0f2-452b-bcba-4c88e055c3ca