Added support for code generation of array references if indexed by indices of Index Type. Specialized for tuple and integer.
Added support for code generation of for loop  statements to
allow indices of IndexType to be used.
Bypass initialization of index variables of IndexType. The
problem seems to be that tuples have no type symbol in
indices which happen to have tuple type. I'll wait on that
until after tuples are more stable.

Added support for index variables of Index Type in array
initialization.

Added two new tests to test the functionality of index type
insertion. These are made futures since they fail with the
'dtInteger' version. To get these working one must change the
flag in symtab.h (_dtinteger_indexType_switch), as well as
uncomment preProcessStmt in insertIndexType.h and
insertIndexType.cpp.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2580 3a8e244f-b0f2-452b-bcba-4c88e055c3ca