Added a new traversal (not a pass) call ClearTypes that sets all the types of defined symbols to dtUnknown.
Made function cloning call the ClearTypes traversal since we will
require analysis to get the types for functions it clones.

Added VarInitExpr to the UpdateSymbols traversal so that copying it
will update its symbol if that changed.  This doesn't yet seem to help
test_typevar_class1.chpl, but I'll look at it further.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2648 3a8e244f-b0f2-452b-bcba-4c88e055c3ca