Added code to remove dead symbols from their scopes and to remove the ParamScopes of dead functions in RemoveDeadSymbols.
Fixed a bug in the Symboltable traversal mechanism that showed up now
that symbols are removed from their scopes within RemoveDeadSymbols.

Fixed a small inconsistency in InsertThisParameters in which a DefStmt
was copied and inserted before being extracted.  I'm not checking in
my code that caught this inconsistency as I don't think it is a
worthwhile inconsistency to catch.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2659 3a8e244f-b0f2-452b-bcba-4c88e055c3ca