Reworked ScopeResolveSymbols to work on methods in classes even though they are no longer scoped within the class.  This is a major change to this traversal.
Added functionality to InsertThisParameters to move methods out of
classes.  I did this before, but now it is done right, or at least,
righter.

Reworked ApplyThisParameters to not use the scope to determine if a
symbol needs 'this' applied to it.  This can stand to be thought about
more and maybe moved into ScopeResolveSymbols.

Fixed OmitForNoAnalysis to actually remove symbols that are omitted,
not just mangle their name.

Added code to copy and update the _this field of a function if the
function is copied in UpdateSymbols (called by copy).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2674 3a8e244f-b0f2-452b-bcba-4c88e055c3ca