Added new AST expression CastLikeExpr that is similar to a CastExpr but instead of a Type* newType, it has Variable* variable.  This casts the expression to the type of the variable.
This is useful in malloc, because when the constructor is cloned, we
want the type associated with the malloc to change.

I'm a little concerned about casts and cloning though.  When a
function is cloned, should the cast be updated?  It seems like
sometimes it should be.  If we can get this working, I can undo my
last changes to SizeofExpr and this CastLikeExpr.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2698 3a8e244f-b0f2-452b-bcba-4c88e055c3ca