In a discussion about portability between gcc on Sun vs. everything else, realized that I'm increasingly offended by our need to #include inttypes.h, and so came up with a hack that allows us to use stdint.h (which also supports other abilities to specialize for platforms that we might anticipate needing).
Added the ability to add a $PLATFORM-specific
subdirectory to the include/ directories in
the compiler and runtime, which will be
searched prior to the main include directory,
allowing platform-specific include files to
be created.  Instantiated this directory for
sunos, and put in a stdint.h file that
#includes inttypes.h.  Changed all #inclusions
of inttypes.h to stdint.h.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2702 3a8e244f-b0f2-452b-bcba-4c88e055c3ca