This checkin gets class cloning working!  At least on the simple test cases that are written.
Changed VarInitExpr to have an Expr rather than a VarSymbol.  I may
change it back.  Note that this is only and can only be used for the
default arguments to the default class constructor.

Inserted code after resolveTypes in a new traversal called
PostAnalysisCleanup that resolves the types of parameters to default
constructors (and to default wrappers as they are copied).

Fixed a bug in which the wrapper did not create ParenOpExprs but
rather FnCalls so they were not resolved.  The coercion wrapper still
creates a FnCall because it looks like call_info works wrong on it.

Added new class cloning test

  test/types/type_variables/deitz/test_clone_class2b

as a simple test where the constructors are passed arguments.

Removed futures for

  test/types/type_variables/deitz/test_clone_class2
  test/types/type_variables/deitz/test_clone_class2a


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2712 3a8e244f-b0f2-452b-bcba-4c88e055c3ca