Added a new pass called Verify.  This replaces uses of Fixup-verify in the passlist.  It calls Fixup-verify.
Fixup-verify only checked the parent pointers.  Verify will eventually
check alot more.

In addition to future verify checks for every kind of expression and
statement in a verify method, this verify pass checks to make sure all
the symbols in a Symboltable traversal are also define in an AST
traversal.  I was running into huge problems before because things
showed up in one kind of traversal and not the other.

This test is currently off because IndexTypes are not fully defined.

Stopped extracting the DefStmt in InsertVariableInitializations so
that the verify continues to work.  Maybe this should be done by
changing what is verified in the passlist.

I added the CondStmt::verify and DefExpr::verify procedures but
eventually we want verify procedures in all statements and
expressions.  As an example of what these will do, the CondStmt makes
sure that condExpr and thenStmt are non-null and makes sure that
condExpr, thenStmt, and elseStmt are single expressions/statements
rather than lists.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2737 3a8e244f-b0f2-452b-bcba-4c88e055c3ca