Modified Symboltable::lookup etc. to lookup symbols in the class type's scope at the appropriate point when looking up symbols in methods.  The appropriate point is after the symbols in the parameter list of the method are looked at.
Simplified ScopeResolveSymbols by removing code to resolve symbols in
methods since this is now in lookup.

Added a new kind of symbol called a ForwardingSymbol.  These symbols
will eventually be included in scopes that use the "use" keyword.  The
plan is that in the AST, the Variable* will refer directly to the
symbol that it is using.  It will also contain a ForwardingSymbol
field that will be used in code generation if the symbol was renamed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2752 3a8e244f-b0f2-452b-bcba-4c88e055c3ca