Added initial support for chapel programs with multiple modules that make "use" of the multiple modules.  Added two new tests, which work, that test this new functionality slightly.
In making this change, I made the default export true.  I would like
to discuss with Brad an option of having a single header file for
symbols that are exported.  Given the forwarding symbol and limited
code mangling, I think we can get away with this.

Added a new pass ResolveModuleUses that adds forwarding symbols to
scopes where a use is used.  It also adds the module that is used to a
list of modules in the current module so that the module will #include
the other.  In addition, this pass adds a function call to the init
function of that module to the init function of the current module.
Perhaps it should be called at the site of the use instead.

Added a forwarding symbol to variable.  I think this is unnecessary
unless we want to change the name of forwarded symbols.  While that
might be nice for code generation I think it will be needlessly
complicated.

Added a function Symboltable::getCurrentModule() that returns the
current module during a traversal.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2758 3a8e244f-b0f2-452b-bcba-4c88e055c3ca