Improved InsertFunctionTemps so that it doesn't insert a function temp if the scope of the statement is different than the scope of the function call.  This is the case, for example, when a function call is in a let expression.  Also fixed a bug in InsertFunctionTemps in which it did not insert function temps for function calls in the initialization expressions to DefExprs.
Made TransformLetExprs call InsertFunctionTemps on the block statement
it creates (since it will not have applied to this code before).

Improved and rewrote TransformLetExprs to create a new scope and copy
the symbols deeply.  Previously I just stole the let expression scope
and stuck it in the block statement.  This new way is more robust.  It
was difficult before because copying and replacing weren't as good as
they are now.

Added a test of 'record mycomplex' which will eventually be the code
we use for complex numbers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2770 3a8e244f-b0f2-452b-bcba-4c88e055c3ca