Fix test_class_list1.future.
The issue was how to handle dispatch on "nil".  The problem being
that the default constructor has a "nil" argument.

class node {
  var next: node;
}

the constructor is :

function construct_node(var x : node) ...

And the call from the default wrapper is:

function defaulted_node() {
  return construct_node(next = nil);
}

This will currently fail, because nil is not a subtype of node.

There were two solutions, one of which would have been to remove the
type declaration for the constructor parameter.

Solution 1:

function construct_node(var x)  ...

The second solution to use a "aspect" (a statement to
the compiler to use a particular type for dispatch purposes).

Solution 2:

function defaulted_node() {
  return construct_node(next = nil:node);
}

The "aspect" tells the compiler what type to dispatch to for the nil.

Generally, programmer's should be aware that they cannot dispatch on nil,
and we might want to provide a aspects, at least of nil, as a part of the
language.

I implemented Solution 2 because it was the least intrusive.

We might also want to consider also using Solution 1.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2786 3a8e244f-b0f2-452b-bcba-4c88e055c3ca