See note below about what is commented out.  Complex is still handled the old way.
Added standard include module _chpl_complex.chpl in runtime/include/.

This is now parsed in.  It is labeled as a MOD_STANDARD to distinguish
it from MOD_USER.

Traversals can now be run on the following set of modules:

  MODULES_USER - only user modules
  MODULES_CODEGEN - user modules, common module, standard modules
  MODULES_ALL - user, common, standard, and internal

Traversals that are restricted now set whichModules to MODULES_CODEGEN
instead of MODULES_COMMON_AND_USER.

Added a pragma for renaming a symbol.  So in the complex record, we
use a pragma to rename it (cname) to _chpl_complex because complex
clashes with the C compiler's complex.

Added a pass to change complex literals to complex record constructor
invocations.

NOTE: This pass is not invoked currently.  Also, the complex module is
not used by other modules.  The code to do these is commented out.
The main first problem is that read and write need to be made more
robust.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2813 3a8e244f-b0f2-452b-bcba-4c88e055c3ca