Chaned Roxana's future tests that were working under certain conditions when she checked her code in on 4/28/05 to use the .compopts feature with the new --tmp flag so that they will be tested on a nightly basis.
Re-enabled these tests by putting in yet
another conditional in resolveTypes that
determines whether or not a given type
should be replaced with the output of
analysis.  In particular, made IndexTypes
not get replaced by analysis because it
seems to replace them with integers.

Hoisted the bulk of this conditional into
a helper function replaceTypeWithAnalysisType()
in an attempt to make the code a bit more
readable.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2819 3a8e244f-b0f2-452b-bcba-4c88e055c3ca