Made FnSymbol::instantiate_generic instantiate all the methods of a class when it is called on the constructor of that class.  It does not copy other functions that act on the instantiated class yet.
I also made it so that the Map<BaseAST*,BaseAST*>* parameter of
instantiate_generic is filled in.  This keeps track of everything that
is copied when instantiate_generic is called.

Fixed a bug in the copy infrastructure in which copy rather than
copyInternal was called in VarInitExpr::copy.  Also I extended
copyInternal to work on Symbols and Types because these should call
the copyInternals of expressions and statements as well.

Improved UpdateSymbols to update MemberAccesses.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2835 3a8e244f-b0f2-452b-bcba-4c88e055c3ca