Changed implementation of argument types "F" and "f" in arg.cpp to assume that the target is a bool rather than an int and converted all such target variables from ints to bools.
This change was motivated by a subtle bug that showed
up on platforms with a different endianness (like sunos),
in which accidentally declaring the target as a bool
rather than an int would cause it to not be set
properly.  By making arg.cpp assume a bool target,
the reverse mistake (using an int that's used only as
a boolean value) should not result in a similar error
since integer tests are for zero vs. nonzero and ints
are bound to be at least as large as ints.  Moreover,
it seems reasonable to use bool values to represent
boolean toggles.

Also, removed an extraneous linefeed from the
insertAnonymousTypes pass that only showed up for
Roxana's tests.  Updated the output of these tests
to get rid of the linefeed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2850 3a8e244f-b0f2-452b-bcba-4c88e055c3ca