This checkin implements write and writeln using overloaded functions.
SpecializeParenOpExprs, though this should eventually be removed, now
implements the transformation:

  writeln(x, y, z);  ==>  write(x); write(y); write(z); writeln();

The write functions are now free to be treated like functions resolved
via function overloading.  We implement the primitive write functions
in the prelude for integer, float, string, boole, complex.

Put in special stopgap write functions for unions, arrays, enums,
domains, and sequences.  These stopgaps are found at code generation
and specialized code is generated to handle these objects.  Eventually
this will be cleaned up and these stopgaps removed.

Removed the call to the pass InsertWriteFns.  These are now inserted
in BuildClassConstructorsEtc.  This pass may eventually be a candidate
for making into multiple passes or at least renaming.  Maybe something
like BuildDefaultFunctionsOnTypes.

This checkin contains a patch from John to create __primitives for
write, read and tostring in the preludes.

Note that write is overloaded over nilType as well.  This may or may
not be necessary.  It is currently done so that we can print out the
value of an object of a class that is never assigned.

With this checkin arrays of floats and complexes of floats now print
out correctly (.0).

This checkin updates some of Shannon's memtests because the memory
usage has changed.  Other .goods have changed too.  Various reasons.
New analysis errors because write is a function, new flowanalysis
output.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2899 3a8e244f-b0f2-452b-bcba-4c88e055c3ca