Remembered that type variables in parameter lists to functions should be ParamSymbols not TypeSymbols.  I made the parameter to constructors for classes with type variables take a TypeSymbol.
For discussion, and I think John wanted to do this, but we should
probably have another Type subtype that could point to a TypeSymbol
that is a VariableType.  I just use the TypeSymbol that is in the
class currently for the constructor, but when there are more
constructors, we may need multiple constructors with the type
parameter.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2901 3a8e244f-b0f2-452b-bcba-4c88e055c3ca