Greatly improved line number and filename information in the compiler.
Also, if some of the AST is built after parsing in a traversal, the
name of the traversal is stored in the BaseAST under the field
traversalInfo.

I also added a field copyInfo that is not yet used.  This will tell in
what traversal the ast was copied.  It is a vector of strings since it
could happen in multiple traversals.

I think the traversalInfo should be a vector of strings too since new
ast nodes could be constructed based on new ast nodes.

If the developer flag is set, the errors now print out information
about the traversals where the thing that led to an error was created.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2933 3a8e244f-b0f2-452b-bcba-4c88e055c3ca