Known regression:  trivial/deitz/test_format_primitives.chpl
This checkin switches complex to the new class based implementation
and enables better support for modules.

Made two subdirectories in the new top-level modules directory called
standard and provided.  These contains standard modules that are
included in all Chapel programs and provided modules that the Chapel
programmer can include with a use-statement.  Moved _chpl_complex.chpl
into modules/standard.

Added hacks to maintain reading complexes, casting complexes to
strings, and setting complex config vars.

Made ProcessParameters set the noDefaultInit flag of its temps.  This
is necessary when its temps are of class types and the constructor
would have need to be resolved by analysis.  I'm not sure why the
defaultConstructor is not set actually.  This should be examined to
see if it is necessary or why the defaultConstructor is not right.

Added a "no codegen" pragma for functions that are in the runtime and
don't need to be codegened.  Of course this is the default with the
preludes, but for the complex read and tostring functions, they don't
need to be generated but can't be in the prelude for scoping reasons.

Added a vec of SymScopes to every module to locate where the module
was used.  This is important because if after the uses are resolved,
something is added to the top-level of a module, it should be included
into that module.  This is now so.

Enabled ConstructComplexLiterals to change complex literals into calls
to the complex constructor.

Removed findModuleScope since getModule is just as good.  Beefed up
getModule function.

Futurized classes/deitz/test_mycomplex_record since the complex is the
same.  It dawns on me that complex should be a class so that it is
nominally typed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2937 3a8e244f-b0f2-452b-bcba-4c88e055c3ca