Fixed a bug in type cloning in which an infinite recursion would be encountered if the type was recursive.
Reworked the overloading code so that anything can be overloaded.
ScopeResolveSymbols now checks to make sure only functions are
overloaded.  After this, the compiler can overload other things.  This
is useful especially for type cloning.  I think we can also clean up
the getter so that it is just a part of overloading, overloaded with
the field.

In ApplyThisParameters, this is no longer applied to types.

Made some superficial changes to sequence handling.  This will all
disappear soon.

Modified the error for redefinition of variables to print out all
redefinitions not just one.

Removed futures for test_myseq0,1,2.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2987 3a8e244f-b0f2-452b-bcba-4c88e055c3ca