Improved ScopeResolveSymbols to handle more complex unresolved types. For example, given a class foo that contains a type t, foo.t is now an unresolved type.  Modified UnresolvedType so that it now contains a Vec<char*>* which contains the names for looking up that type.  For foo.t, the strings foo and t would be stored in the UnresolvedType in that order.
Potential concern:

  class foo {
    var x : integer = 12;
    type t : integer = x;
  }

  var k : foo.t;

The problem is that the default value of the type in the class is
dependent on an instance of that class.  We must disallow foo.t in
this case.

Related, are we supporting something like f.t if f is an instance of
foo.  The problem with this is that we could not easily resolve the
type symbol in ScopeResolveSymbols, i.e., before analysis.  I assume
were not allowing something like that.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@2993 3a8e244f-b0f2-452b-bcba-4c88e055c3ca