Implemented sequences using the seq2 class in the _chpl_seq standard module.
Sequence literals are dealt with before analysis, for now.  So they
can only be created if the type of the elements are known.

Sequence class types are created in InsertAnonymousTypes.

Added typeInfo to symbols which should be used to get the type.  This
handles ForwardingSymbols from module uses better.

Removed noanalysis sequence tests and some of the support code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3018 3a8e244f-b0f2-452b-bcba-4c88e055c3ca