Implemented non-analysis support for user-defined default constructors.
Added constructor keyword to grammar.  Added ability to specify a
user-defined default constructor by calling the constructor
'initialize' though I would like to change it to be the class name
instead.

The constructor is modifed in InsertThisParameters where 'this' is
declared at the top of the constructor and returned at the end.  The
normal default constructor is not built for a class if the user
already specified one.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3056 3a8e244f-b0f2-452b-bcba-4c88e055c3ca