Collapsed Loc:: ILink:: BaseAST:: hierarchy.  There is no longer Loc or ILink, just BaseAST.  This allows us to clean up the code in some non-superficial ways.  Moved prev, next, filename, and lineno into BaseAST.
Moved parentScope out of Symbol, Stmt, and Expr, and into BaseAST.  It
is not currently used in Type or AList.

Changed call_fixup to just fixup and made it more focused.  This
greatly improves the performance of compilation.  For example,
SpecializeParenOps on trivial/shannon/compSampler now takes 0.056
seconds on my machine rather than 12.87 seconds!

Unified remove, insertBefore, insertAfter, etc.  The current functions
can now be called when the nodes are in the AST and when they are not.
If they are, fixup is called, otherwise it is not.

  BaseAST::insertBefore(BaseAST* new_ast) inserts new_ast before this
    in a list

  BaseAST::insertAfter(BaseAST* new_ast) inserts new_ast after this in
    a list

  BaseAST::replace(BaseAST* new_ast) replaces this with new_ast

  BaseAST::remove() removes this from the AST, remove does not yet
    take into account symbols and removing them from the symboltable

The functions on AList now include

  AList<...>::insertAtTail(elem) inserts elem at the end of the list

  AList<...>::insertAtHead(elem) inserts elem at the beginning of the
    list

  AList<...>::add(alist) inserts the contents of a list at the end of
    the list


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3108 3a8e244f-b0f2-452b-bcba-4c88e055c3ca