Added support for the alternative syntax for anonymous record types:
  var r : (x : integer, y : string);

This syntax requires that the type be specified, otherwise it is
indistinguishable from a tuple type.

Eliminated DefStmt::typeDef() and DefStmt::moduleDef().  Started to
remove DefStmt::fnDef() and DefStmt::varDef().  These are bad
functions because they only return a single symbol though multiple may
exist.  Added functions DefStmt::definesTypes, definesFunctions, and
definesVariables that return true if the DefStmt only defines those
symbols.

Removed the future for test_anon_record3 and added a new test that
declares two fields in a class in the same line.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3126 3a8e244f-b0f2-452b-bcba-4c88e055c3ca