Removed support for # used to append or prepend to a sequence.  This operator can only be used for sequence concatenation now.  Added hack to support prepending and appending integers and strings since we can tell they are not sequences before analysis.
Cleaned up support for sequences.

Readded a test in insertBefore, insertAfter, and replace to make sure
that what is being added is not in a list already (and thus needs to
be copied).  Maybe this should also check the parents (since then it
would need to be copied as well).

Removed four more sequence futures.  The current two sequence failures
are caused by:

  1. not copying a sequence that is appended/prepended/concatenated to
     and is not dead.

  2. iterating over a sequence that is not a sequence of integers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3174 3a8e244f-b0f2-452b-bcba-4c88e055c3ca